+++ CHANGELOG.md
@@ -14,4 +14,6 @@ and **Merged pull requests**. Critical items to know are:
 The versions coincide with releases on pip. Only major versions will be released as tags on Github.
 
 ## [0.0.x](https://github.com/flux-framework/flux-operator/tree/main) (0.0.x)
+ - Support for automated testing of examples/tests (0.0.x)
+  - Addition of local volumes for workflows that share data.
  - Project skeleton release (0.0.0)